requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
nltk==3.8.1
pyttsx3==2.90
transformers 
torch 
requests
sentencepiece


-e.